livekit-server/requirements.txt
livekit-api>=0.6.0
flask>=3.0.0
flask-cors>=4.0.0

